+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* `DocCfg` new property `num_preview_chars`, standardize preview length computation
 * Include author in `Document.people()` even if `is_valid_for_name_scan` is false
 * Coerce all timestamps to UTC timezone instead of no timezone